pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portreaper"
version = "0.1.0"
description = "The Grim Reaper for processes hogging your ports 💀"
authors = [{name = "Your Name", email = "you@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "psutil>=5.9",
    "python-nmap>=0.7.1"
]

[project.urls]
Homepage = "https://github.com/yourusername/portreaper"
Repository = "https://github.com/yourusername/portreaper"

[project.scripts]
portreaper = "portreaper.main:cli"
